mod authority contract

See merge request !99

+++ _contracts/contracts/Constants.sol
@@ -2,18 +2,18 @@ pragma solidity ^0.4.18;
 
 library Constants {
     function params() pure internal returns (address) {
-        return address(bytes2("pa"));
+        return address(uint16(bytes2("pa")));
     }
 
     function energy() pure internal returns (address) {
-        return address(bytes2("en"));
+        return address(uint16(bytes2("en")));
     }
 
     function authority() pure internal returns (address) {
-        return address(bytes2("au"));
+        return address(uint16(bytes2("au")));
     }
 
     function god() pure internal returns (address) {        
-        return address(bytes2("go"));
+        return address(uint16(bytes2("go")));
     }
 }
